[Engine] Many warnings removed + issue with selection + issue with async model loading

+++ Source/Framework/Core/Animation/TeAnimation.cpp
@@ -112,7 +112,7 @@ namespace te
                     tempLayers.push_back(AnimationStateLayer());
                     AnimationStateLayer& newLayer = tempLayers.back();
 
-                    newLayer.Index = layer;
+                    newLayer.Index = static_cast<UINT8>(layer);
                 }
 
                 clipIdx++;
